feat: more clippy fixes

+++ src/mem.rs
@@ -279,7 +279,7 @@ impl Compress {
 
         match rc {
             ffi::MZ_STREAM_ERROR => compress_failed(self.inner.inner.msg()),
-            ffi::MZ_OK => Ok(unsafe { (*stream).adler } as u32),
+            ffi::MZ_OK => Ok(unsafe { (*stream).adler }),
             c => panic!("unknown return code: {}", c),
         }
     }
@@ -497,8 +497,8 @@ impl Decompress {
 
         match rc {
             ffi::MZ_STREAM_ERROR => decompress_failed(self.inner.inner.msg()),
-            ffi::MZ_DATA_ERROR => decompress_need_dict(unsafe { (*stream).adler } as u32),
-            ffi::MZ_OK => Ok(unsafe { (*stream).adler } as u32),
+            ffi::MZ_DATA_ERROR => decompress_need_dict(unsafe { (*stream).adler }),
+            ffi::MZ_OK => Ok(unsafe { (*stream).adler }),
             c => panic!("unknown return code: {}", c),
         }
     }
